% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-run.R
\name{app_server}
\alias{app_server}
\title{Wire the stage module servers around a shared meta store}
\usage{
app_server(input, output, session)
}
\description{
Wire the stage module servers around a shared meta store
}
\keyword{internal}
