% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-binary.R
\name{binary_server}
\alias{binary_server}
\title{Binary Mixture stage server}
\usage{
binary_server(id, meta)
}
\arguments{
\item{id}{Module id.}

\item{meta}{Shared reactiveValues for experiment metadata.}
}
\description{
Binary Mixture stage server
}
\keyword{internal}
