% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-binary.R
\name{bound_row}
\alias{bound_row}
\title{One lower/upper bound numeric-input pair for a parameter}
\usage{
bound_row(ns, param, label, hi_default = NA)
}
\description{
One lower/upper bound numeric-input pair for a parameter
}
\keyword{internal}
