% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-ternary-asa.R
\name{ca_asa_tri}
\alias{ca_asa_tri}
\title{ca_asa_tri: ternary Advanced S/A CA predictor}
\usage{
ca_asa_tri(
  c1,
  c2,
  c3,
  max,
  slope1,
  slope2,
  slope3,
  ec50_1,
  ec50_2,
  ec50_3,
  A1,
  A2,
  A3,
  A4
)
}
\arguments{
\item{c1, c2, c3}{Concentrations.}

\item{max, slope1, slope2, slope3, ec50_1, ec50_2, ec50_3}{Base curve parameters.}

\item{A1, A2, A3}{Pairwise interaction coefficients (A1: C1-C2, A2: C1-C3,
A3: C2-C3).}

\item{A4}{Three-way interaction coefficient.}
}
\value{
Predicted response (scalar).
}
\description{
ca_asa_tri: ternary Advanced S/A CA predictor
}
\keyword{internal}
