% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-binary.R
\name{ca_bi}
\alias{ca_bi}
\title{ca_bi: binary (2-chemical) mixture predictor (ported verbatim from CA_bi)}
\usage{
ca_bi(c1, c2, max, slope1, slope2, ec501, ec502)
}
\arguments{
\item{c1, c2}{Concentrations.}

\item{max, slope1, slope2, ec501, ec502}{Model parameters.}
}
\value{
Predicted response (scalar).
}
\description{
ca_bi: binary (2-chemical) mixture predictor (ported verbatim from CA_bi)
}
\keyword{internal}
