% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-ternary.R
\name{ca_dr_tri}
\alias{ca_dr_tri}
\title{ca_dr_tri: ternary (3-chemical) mixture predictor (ported verbatim from CA_DR)}
\usage{
ca_dr_tri(
  c1,
  c2,
  c3,
  max,
  slope1,
  slope2,
  slope3,
  ec50_1,
  ec50_2,
  ec50_3,
  a,
  b1,
  b2,
  b3
)
}
\arguments{
\item{c1, c2, c3}{Concentrations.}

\item{max, slope1, slope2, slope3, ec50_1, ec50_2, ec50_3, a, b1, b2, b3}{Model parameters.}
}
\value{
Predicted response (scalar).
}
\description{
ca_dr_tri: ternary (3-chemical) mixture predictor (ported verbatim from CA_DR)
}
\keyword{internal}
