% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-binary.R
\name{ca_sa_bi}
\alias{ca_sa_bi}
\title{ca_sa_bi: binary (2-chemical) mixture predictor (ported verbatim from CA_SA_bi)}
\usage{
ca_sa_bi(c1, c2, max, slope1, slope2, ec501, ec502, a)
}
\arguments{
\item{c1, c2}{Concentrations.}

\item{max, slope1, slope2, ec501, ec502, a}{Model parameters.}
}
\value{
Predicted response (scalar).
}
\description{
ca_sa_bi: binary (2-chemical) mixture predictor (ported verbatim from CA_SA_bi)
}
\keyword{internal}
