% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-ternary.R
\name{ca_sa_tri}
\alias{ca_sa_tri}
\title{ca_sa_tri: ternary (3-chemical) mixture predictor (ported verbatim from CA_SA)}
\usage{
ca_sa_tri(c1, c2, c3, max, slope1, slope2, slope3, ec50_1, ec50_2, ec50_3, a)
}
\arguments{
\item{c1, c2, c3}{Concentrations.}

\item{max, slope1, slope2, slope3, ec50_1, ec50_2, ec50_3, a}{Model parameters.}
}
\value{
Predicted response (scalar).
}
\description{
ca_sa_tri: ternary (3-chemical) mixture predictor (ported verbatim from CA_SA)
}
\keyword{internal}
