% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-campaign.R
\name{campaign_pairs}
\alias{campaign_pairs}
\title{The pairs of a campaign, in A-B, A-C, B-C order}
\usage{
campaign_pairs(chems)
}
\arguments{
\item{chems}{Integer vector of stressor indices.}
}
\value{
A list of length-2 integer vectors.
}
\description{
The pairs of a campaign, in A-B, A-C, B-C order
}
\keyword{internal}
