% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-curve-fit.R
\name{curve_fit_ui}
\alias{curve_fit_ui}
\title{Curve-fit panel UI (plots + parameter grid + Autofit/Simulate)}
\usage{
curve_fit_ui(id)
}
\arguments{
\item{id}{Module id.}
}
\description{
Curve-fit panel UI (plots + parameter grid + Autofit/Simulate)
}
\keyword{internal}
