% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternary-isoplane.R
\name{.f4}
\alias{.f4}
\title{F4 deviation factor for a ternary Advanced-S/A model}
\usage{
.f4(z1, z2, z3, A1, A2, A3, A4)
}
\description{
F4 deviation factor for a ternary Advanced-S/A model
}
\keyword{internal}
