% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-binary.R
\name{ia_bi}
\alias{ia_bi}
\title{ia_bi: binary (2-chemical) mixture predictor (ported verbatim from IA_bi)}
\usage{
ia_bi(c1, c2, max, slope1, slope2, ec501, ec502)
}
\arguments{
\item{c1, c2}{Concentrations.}

\item{max, slope1, slope2, ec501, ec502}{Model parameters.}
}
\value{
Predicted response (scalar).
}
\description{
ia_bi: binary (2-chemical) mixture predictor (ported verbatim from IA_bi)
}
\keyword{internal}
