% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-binary.R
\name{ia_dr_bi}
\alias{ia_dr_bi}
\title{ia_dr_bi: binary (2-chemical) mixture predictor (ported verbatim from IA_DR_bi)}
\usage{
ia_dr_bi(c1, c2, max, slope1, slope2, ec501, ec502, a, b)
}
\arguments{
\item{c1, c2}{Concentrations.}

\item{max, slope1, slope2, ec501, ec502, a, b}{Model parameters.}
}
\value{
Predicted response (scalar).
}
\description{
ia_dr_bi: binary (2-chemical) mixture predictor (ported verbatim from IA_DR_bi)
}
\keyword{internal}
