% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-binary.R
\name{ia_sa_bi}
\alias{ia_sa_bi}
\title{ia_sa_bi: binary (2-chemical) mixture predictor (ported verbatim from IA_SA_bi)}
\usage{
ia_sa_bi(c1, c2, max, slope1, slope2, ec501, ec502, a)
}
\arguments{
\item{c1, c2}{Concentrations.}

\item{max, slope1, slope2, ec501, ec502, a}{Model parameters.}
}
\value{
Predicted response (scalar).
}
\description{
ia_sa_bi: binary (2-chemical) mixture predictor (ported verbatim from IA_SA_bi)
}
\keyword{internal}
