% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models-ternary.R
\name{ia_tri}
\alias{ia_tri}
\title{ia_tri: ternary (3-chemical) mixture predictor (ported verbatim from IA)}
\usage{
ia_tri(c1, c2, c3, max, slope1, slope2, slope3, ec50_1, ec50_2, ec50_3)
}
\arguments{
\item{c1, c2, c3}{Concentrations.}

\item{max, slope1, slope2, slope3, ec50_1, ec50_2, ec50_3}{Model parameters.}
}
\value{
Predicted response (scalar).
}
\description{
ia_tri: ternary (3-chemical) mixture predictor (ported verbatim from IA)
}
\keyword{internal}
