% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-intro.R
\name{intro_server}
\alias{intro_server}
\title{Introduction stage server}
\usage{
intro_server(id, meta)
}
\arguments{
\item{id}{Module id.}

\item{meta}{Shared reactiveValues for experiment metadata.}
}
\description{
Introduction stage server
}
\keyword{internal}
