% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{lr_test}
\alias{lr_test}
\title{Likelihood-ratio test between a model and its nesting parent}
\usage{
lr_test(obj_parent, obj_child, df_parent, df_child, n, response)
}
\arguments{
\item{obj_parent, obj_child}{Objective values (SS for continuous, deviance for
binary) of the simpler (parent) and more complex (child) models.}

\item{df_parent, df_child}{Number of free parameters in each model.}

\item{n}{Number of observations (used for the continuous statistic).}

\item{response}{"continuous" or "binary".}
}
\value{
A list with \code{chi}, \code{df}, and \code{p} (upper-tail chi-squared p-value).
}
\description{
Likelihood-ratio test between a model and its nesting parent
}
