% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objective.R
\name{obj_ss}
\alias{obj_ss}
\title{Sum-of-squares objective for continuous responses}
\usage{
obj_ss(obs, pred)
}
\arguments{
\item{obs, pred}{Numeric vectors of observed and predicted responses.}
}
\value{
The residual sum of squares.
}
\description{
Sum-of-squares objective for continuous responses
}
