% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-campaign.R
\name{pair_key}
\alias{pair_key}
\title{Store key for a pair}
\usage{
pair_key(i, j)
}
\arguments{
\item{i, j}{Stressor indices.}
}
\value{
A character key, e.g. \code{"23"}.
}
\description{
Store key for a pair
}
\keyword{internal}
