% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-io.R
\name{read_upload}
\alias{read_upload}
\title{Read an uploaded CSV file}
\usage{
read_upload(path)
}
\arguments{
\item{path}{File path.}
}
\value{
A data frame.
}
\description{
Read an uploaded CSV file
}
\keyword{internal}
