% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-singles.R
\name{singles_server}
\alias{singles_server}
\title{Singles stage server}
\usage{
singles_server(id, store)
}
\arguments{
\item{id}{Module id.}

\item{store}{The campaign store.}
}
\description{
Singles stage server
}
\keyword{internal}
