% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixdra-package.R
\docType{package}
\name{mixdra-package}
\alias{mixdra}
\alias{mixdra-package}
\title{mixdra: Mixture Dose-Response Analysis (Jonker et al. 2005 Methodology)}
\description{
Fits single-chemical and binary/ternary mixture dose-response models using the concentration addition and independent action reference models with synergism/antagonism, dose-ratio-dependent, and dose-level-dependent deviation patterns, following Jonker et al. (2005) \doi{10.1897/04-431R.1}. Supports continuous and binary (quantal) responses and likelihood-ratio model comparison.
}
\author{
\strong{Maintainer}: Mixdra Authors \email{pieter.vklaveren@gmail.com}

Authors:
\itemize{
  \item Mixdra Authors \email{pieter.vklaveren@gmail.com}
}

}
\keyword{internal}
