% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-curve-fit.R
\name{single_model_equation}
\alias{single_model_equation}
\title{Visible model-equation header for the curve-fit panel}
\usage{
single_model_equation()
}
\value{
A \code{shiny} tag.
}
\description{
Visible model-equation header for the curve-fit panel
}
\keyword{internal}
