% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-single.R
\name{single_server}
\alias{single_server}
\title{Single Stressor stage server}
\usage{
single_server(id, meta)
}
\arguments{
\item{id}{Module id.}

\item{meta}{Shared reactiveValues for experiment metadata.}
}
\description{
Single Stressor stage server
}
\keyword{internal}
