% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-ternary.R
\name{ternary_server}
\alias{ternary_server}
\title{Ternary Mixture stage server}
\usage{
ternary_server(id, meta)
}
\arguments{
\item{id}{Module id.}

\item{meta}{Shared reactiveValues for experiment metadata.}
}
\description{
Ternary Mixture stage server
}
\keyword{internal}
