% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-io.R
\name{upload_schema}
\alias{upload_schema}
\title{Fixed column schema for a stage and response type}
\usage{
upload_schema(stage, response)
}
\arguments{
\item{stage}{"single" or "binary".}

\item{response}{"continuous" or "quantal".}
}
\value{
Character vector of required column names.
}
\description{
Fixed column schema for a stage and response type
}
\keyword{internal}
